Disable caching of FreeIPA client
Tokens issued by calling the Client.login() method have a TTL, which means once the TTL is reached the cached client is no longer valid.

This can be reintroduced in future if we can find a way to discover the validity of tokens or their TTL and use a time-based cache.